Particle emitter: 3-mode flow field system + highp precision
Replace the broken backward-trace field-flow approach with a proper
3-mode architecture:

• LINEAR: existing spawn-point mode (particlePos, angle/spread/wind)
• NOISE: forward Euler integration through internal curl/FBM noise field
  — 6 Euler steps, pfbm + flowAngle helpers, no external field node needed
  — seeds distributed in g_uv space around posVar with spawn_radius
• GRAVITY: backward trace via field socket, seeds fixed in g_uv coordinate
  space (fixes the [0,1]×[0,1] vs g_uv space mismatch that made particles
  invisible in all field-flow examples)

Other changes:
- precision mediump → highp to avoid 1e9 overflow in particle timestamps
- particleFlowDrift + particleSpiralVortex → noise mode, external VectorField
  / SpiralField nodes removed (simpler graphs, better visuals)
- particleOrbitCloud + particleMouseAttract → gravity mode with spawn_radius
- NodeInlineViz: mode-aware preview (swirl for noise, scatter for gravity)
- Default posVar: vec2(0.5) → vec2(0.0) (screen centre in g_uv space)

+++ src/compiler/shaderAssembler.ts
@@ -1014,7 +1014,7 @@ export function generateFragmentShader(
     .map(name => `uniform float ${name};`)
     .join('\n');
 
-  const fragmentShader = `precision mediump float;
+  const fragmentShader = `precision highp float;
 #define PI 3.1415926538
 #define TAU 6.2831853072
